% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{advi}
\alias{advi}
\title{Use Automatic Differentiation Variational Inference (ADVI)}
\usage{
advi(bnn, samples_per_step, maxiter)
}
\arguments{
\item{bnn}{A BNN formed using \code{\link{BNN}}}

\item{samples_per_step}{Samples used in each iteration to
estimate ELBO}

\item{maxiter}{Iterations to run AFVI for}
}
\value{
A object of class advi from which can be drawn using
        \code{\link{draw.vi}}
}
\description{
Uses ADVI to approximate the BNN posterior using a diagonal MvNormal
}
