% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{bbb}
\alias{bbb}
\title{Use Bayes By Backprop (BBB)}
\usage{
bbb(bnn, samples_per_step, maxiter, batchsize, nchains = 1)
}
\arguments{
\item{bnn}{A BNN formed using \code{\link{BNN}}}

\item{samples_per_step}{How many samples to take in each iteration (to approximate ELBO)}

\item{maxiter}{Maximum number of iterations to run the algorithm}

\item{batchsize}{Mini-Batchsize to use}

\item{nchains}{How many BBB approximations to form}
}
\value{
A objevt of class bbb from which can be drawn using \code{\link{draw.vi}}
}
\description{
Uses BBB to form a variational approximation to the posterior using
a diagonal multivariate normal distribution.
}
