% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFluxR.R
\name{BFluxR_setup}
\alias{BFluxR_setup}
\title{Set up of the Julia environment needed for BFlux}
\usage{
BFluxR_setup(pkg_check = TRUE, nthreads = 4, seed = NULL, ...)
}
\arguments{
\item{pkg_check}{(Default=TRUE) Check whether needed Julia packages
are installed}

\item{nthreads}{(Default=4) How many threads to make available to Julia}

\item{seed}{Seed to be used.}

\item{...}{Other parameters passed on to \code{\link[JuliaCall]{julia_setup}}}
}
\description{
Set up of the Julia environment needed for BFlux
}
