% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFluxR.R
\name{BNN}
\alias{BNN}
\title{Create A Bayesian Neural Network using BFlux.jl}
\usage{
BNN(net, loglike, y, x)
}
\arguments{
\item{net}{A network created using \code{\link{Chain}}}

\item{loglike}{A likelihood created using, for example, \code{\link{likelihood.feedforward_normal}}}

\item{y}{outcomes}

\item{x}{features}
}
\value{
A list containing the following
\itemize{
    \item juliavar - julia variable containing the BNN
    \item juliacode - julia code used to create the BNN
}
}
\description{
Create A Bayesian Neural Network using BFlux.jl
}
