% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFluxR.R
\name{BNN.totparams}
\alias{BNN.totparams}
\title{Obtain the total parameters of the BNN}
\usage{
BNN.totparams(bnn)
}
\arguments{
\item{bnn}{A BNN formed using \code{\link{BNN}}}
}
\description{
Obtain the total parameters of the BNN
}
