% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFluxR.R
\name{Chain}
\alias{Chain}
\title{Chain various layers together to form a network}
\usage{
Chain(...)
}
\arguments{
\item{...}{Comma separated layers}
}
\value{
List with the following content
\itemize{
    \item juliavar - the julia variable containing the network
    \item specification - the string representation of the network
}
}
\description{
Chain various layers together to form a network
}
