% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFluxR.R
\name{.install_pkg}
\alias{.install_pkg}
\title{Installs Julia packages if needed}
\usage{
.install_pkg(...)
}
\arguments{
\item{...}{strings of package names}
}
\description{
Installs Julia packages if needed
}
