% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFluxR.R
\name{.julia_project_status}
\alias{.julia_project_status}
\title{Obtain the status of the current Julia project}
\usage{
.julia_project_status()
}
\description{
Obtain the status of the current Julia project
}
