% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{draw.laplace}
\alias{draw.laplace}
\title{Draw values from a distribution}
\usage{
draw.laplace(dist, n)
}
\arguments{
\item{dist}{Should either be a prior distribution (see for example
\code{\link{Gamma}}) or a laplace approximation.}

\item{n}{Number of draws}
}
\description{
Draw values from a distribution
}
