% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{draw.vi.advi}
\alias{draw.vi.advi}
\title{Draw from ADVI distribution}
\usage{
\method{draw.vi}{advi}(vi, n)
}
\arguments{
\item{vi}{A variational distribution}

\item{n}{Number of draws}
}
\description{
Draw from ADVI distribution
}
