% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{draw.vi.bbb}
\alias{draw.vi.bbb}
\title{Draw from BBB distribution}
\usage{
\method{draw.vi}{bbb}(vi, n)
}
\arguments{
\item{vi}{A variational distribution}

\item{n}{Number of draws}
}
\description{
Draw from BBB distribution
}
