% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{draw.vi}
\alias{draw.vi}
\title{Draws from a variational distribution}
\usage{
draw.vi(vi, n)
}
\arguments{
\item{vi}{A variational distribution}

\item{n}{Number of draws}
}
\value{
A Matrix; Rows are parameters, columns are draws;
}
\description{
Draws from a variational distribution
}
