% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{find_mode}
\alias{find_mode}
\title{Find the posterior mode of a BNN using full gradients}
\usage{
find_mode(bnn, maxiter, tol = 1e-06)
}
\arguments{
\item{bnn}{BNN created using \code{\link{BNN}}}

\item{maxiter}{maximum iterations}

\item{tol}{Stop when update is less than tol percent}
}
\value{
A list containing
\itemize{
    \item has_converged - Whether the algorithm has converged
    \item mode - The mode (in vector form)
}
}
\description{
Find the posterior mode of a BNN using full gradients
}
