% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{find_mode_sgd}
\alias{find_mode_sgd}
\title{Find the posterior mode of a BNN using Stochastic Gradient Descent}
\usage{
find_mode_sgd(bnn, batchsize, init, maxiter, tol = 1e-06)
}
\arguments{
\item{bnn}{BNN created using \code{\link{BNN}}}

\item{batchsize}{Mini-batchsize}

\item{init}{Vector of intial values; See \code{\link{BNN.totparams}}}

\item{maxiter}{maximum iterations}

\item{tol}{Stop when update is less than tol percent}
}
\value{
A list containing the following
\itemize{
    \item has_converged - Whether the algorithm has converged
    \item mode - The mode (in vector form)
    \item loss - Value of the logposterior at each point along the optimisation
}
}
\description{
Find the posterior mode of a BNN using Stochastic Gradient Descent
}
