% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{Gamma}
\alias{Gamma}
\title{Create a Gamma Prior}
\usage{
Gamma(shape = 2, scale = 2)
}
\arguments{
\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\value{
A list with the following content
\itemize{
    \item juliavar - julia variable containing the distribution
    \item juliacode - julia code used to create the distribution
}
}
\description{
Creates a Gamma prior in Julia using Distributions.jl
}
