% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{ggmc}
\alias{ggmc}
\title{Use Gradient Guided Monte Carlo with Stochastic Gradients}
\usage{
ggmc(
  bnn,
  batchsize,
  maxiter,
  init = NULL,
  nchains = 1,
  l = 0.01,
  beta = 0.5,
  keep_every = 10,
  adapruns = 5000,
  kappa = 0.5,
  goal_accept_rate = 0.65,
  adaptM = FALSE,
  adapth = TRUE
)
}
\arguments{
\item{bnn}{BNN formed using \code{\link{BNN}}}

\item{batchsize}{Mini-Batchsize}

\item{maxiter}{Number of epochs to run ggmc for}

\item{init}{see \code{\link{sgld}}}

\item{nchains}{Number of chains}

\item{l}{Step length}

\item{beta}{Momentum}

\item{keep_every}{If greater than 1, then delayed acceptance will be used.
Only the last draw will be kept.}

\item{adapruns}{How many runs to use for adaptation}

\item{kappa}{Decrease exponent for step size adaptation}

\item{goal_accept_rate}{Target acceptance rate}

\item{adaptM}{Adapt the mass matrix? Often fails currently}

\item{adapth}{Adapt h (This directly corresponds to adapting `l`)}
}
\value{
A list with the following content
\itemize {
    \item samples - A matrix of dimension parameters x draws if `nchains==1`, otherwise
          a tensor of dimensions parameters x draws x chains
    \item hastings - MH ratios.
    \item momenta - Momentum draws
    \item accept - accepted draw = 1
}
}
\description{
Uses GGMC with Stochastic Gradients as proposed in
Garriga-Alonso, A., & Fortuin, V. (2021).
Exact langevin dynamics with stochastic gradients. arXiv preprint arXiv:2102.01691.
}
