% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{InverseGamma}
\alias{InverseGamma}
\title{Create an InverGamma Prior}
\usage{
InverseGamma(shape = 2, scale = 2)
}
\arguments{
\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\description{
Creates and Inverse Gamma prior in Julia using Distributions.jl
}
\seealso{
\code{\link{Gamma}}
}
