% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{laplace}
\alias{laplace}
\title{Form a laplace approximation of the posterior}
\usage{
laplace(bnn, maxiter, M = 1, tol = 1e-06)
}
\arguments{
\item{bnn}{A BNN formed using \code{\link{BNN}}}

\item{maxiter}{Maximum number of itertions to find modes}

\item{M}{Number of modes to find}

\item{tol}{tolerance (converged when subsequence changes are less than `tol` percent apart)}
}
\value{
A list containing
\itemize{
    \item has_converged - Boolean vector giving whether each mode estimate as converged
    \item juliavar - Julia variable holding the laplace approximation
}
}
\description{
Uses `M` mode estimates to form a laplace approximation around
each of these modes. The returned distribution is then a mixture
of normals with equal weights. Currently only diagonal covarainces
are implemented.
}
