% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{likelihood.seqtoone_tdist}
\alias{likelihood.seqtoone_tdist}
\title{Use a T-likelihood for a seq-to-one recurrent network.}
\usage{
likelihood.seqtoone_tdist(sig_prior, nu = 30)
}
\arguments{
\item{sig_prior}{A prior distribution for sigma defined using
\code{\link{Gamma}}, \code{link{InverGamma}},
\code{\link{Truncated}}, \code{\link{Normal}}}

\item{nu}{DF of TDist}
}
\value{
see \code{\link{likelihood.feedforward_normal}}
}
\description{
See \code{\link{likelihood.seqtoone_normal}} and \code{\link{likelihood.feedforward_tdist}}
for details,
}
