% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{LSTM}
\alias{LSTM}
\title{Create an LSTM layer with `in_size` input size, and `out_size` hidden state size}
\usage{
LSTM(in_size, out_size)
}
\arguments{
\item{in_size}{Input size}

\item{out_size}{Output size}
}
\description{
Create an LSTM layer with `in_size` input size, and `out_size` hidden state size
}
\seealso{
\code{\link{Dense}}
}
