% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-mass.R
\name{madapter.FixedMassMatrix}
\alias{madapter.FixedMassMatrix}
\title{Use a fixed mass matrix}
\usage{
madapter.FixedMassMatrix(mat = NULL)
}
\arguments{
\item{mat}{(Default=NULL); inverse mass matrix; If `NULL`, then
identity matrix will be used}
}
\value{
list with `juliavar` and `juliacode` and given matrix or `NULL`
}
\description{
Use a fixed mass matrix
}
