% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-mass.R
\name{madapter.FullCov}
\alias{madapter.FullCov}
\title{Use the full covariance matrix as inverse mass matrix}
\usage{
madapter.FullCov(adapt_steps, windowlength, kappa = 0.5, epsilon = 1e-06)
}
\arguments{
\item{adapt_steps}{Number of adaptation steps}

\item{windowlength}{Lookback window length for calculation of covariance}

\item{kappa}{How much to shrink towards the identity}

\item{epsilon}{Small value to add to diagonal so as to avoid numerical
non-pos-def problem}
}
\value{
see \code{\link{madapter.DiagCov}}
}
\description{
Use the full covariance matrix as inverse mass matrix
}
