% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-mass.R
\name{madapter.RMSProp}
\alias{madapter.RMSProp}
\title{Use RMSProp to adapt the inverse mass matrix.}
\usage{
madapter.RMSProp(adapt_steps, lambda = 1e-05, alpha = 0.99)
}
\arguments{
\item{adapt_steps}{number of adaptation steps}

\item{lambda}{see above paper}

\item{alpha}{see above paper}
}
\value{
list with `juliavar` and `juliacode` and all given arguments
}
\description{
Use RMSProp as a preconditions/mass matrix adapter. This was proposed in Li, C.,
Chen, C., Carlson, D., & Carin, L. (2016, February). Preconditioned stochastic
gradient Langevin dynamics for deep neural networks. In Thirtieth AAAI
Conference on Artificial Intelligence for the use in SGLD and related methods.
}
