% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-finding.R
\name{opt.ADAM}
\alias{opt.ADAM}
\title{ADAM optimiser}
\usage{
opt.ADAM(eta = 0.001, beta = c(0.9, 0.999), eps = 1e-08)
}
\arguments{
\item{eta}{stepsize}

\item{beta}{momentum decays; must be a list of length 2}

\item{eps}{Flux does not document this}
}
\value{
see \code{\link{opt.Descent}}
}
\description{
ADAM optimiser
}
