% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-finding.R
\name{opt.RMSProp}
\alias{opt.RMSProp}
\title{RMSProp optimiser}
\usage{
opt.RMSProp(eta = 0.001, rho = 0.9, eps = 1e-08)
}
\arguments{
\item{eta}{learning rate}

\item{rho}{momentum}

\item{eps}{not documented by Flux}
}
\value{
see \code{\link{opt.Descent}}
}
\description{
RMSProp optimiser
}
