% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFluxR.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Obtain posterior predictive draws}
\usage{
posterior_predict(bnn, samples, newx = NULL)
}
\arguments{
\item{bnn}{A BNN formed using \code{\link{BNN}}}

\item{samples}{Either a matrix of dimensions parameters x draws
or a tensor of dimensions parameters x draws x chains}

\item{newx}{New X to use for posterior predictions}
}
\value{
Returns a matrix of dimensions N x draws if samples is a matrix
        and otherwise a tensor of dimensions N x draws x chains
}
\description{
Obtain posterior predictive draws
}
