% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{posterior_predictive}
\alias{posterior_predictive}
\title{Draw from the posterior predictive distribution}
\usage{
posterior_predictive(bnn, posterior_samples, x = NULL)
}
\arguments{
\item{bnn}{a BNN obtained using \code{link{BNN}}}

\item{posterior_samples}{a vector or matrix containing posterior
samples. This can be obtained using \code{\link{mcmc}}, or \code{\link{bayes_by_backprop}}
or \code{\link{find_mode}}.}

\item{x}{input variables. If `NULL` (default), traning values will be used.}
}
\value{
A matrix whose columns are the posterior predictive draws.
}
\description{
Draw from the posterior predictive distribution
}
