% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior.gaussian}
\alias{prior.gaussian}
\title{Use an isotropic Gaussian prior}
\usage{
prior.gaussian(chain, sigma)
}
\arguments{
\item{chain}{Chain obtained using \code{\link{Chain}}}

\item{sigma}{Standard deviation of Gaussian prior}
}
\value{
a list containing the following
\itemize{
    \item `juliavar` the julia variable used to store the prior
    \item `juliacode` the julia code
}
}
\description{
Use a Multivariate Gaussian prior for all network parameters.
Covariance matrix is set to be equal `sigma * I` with `I` being
the identity matrix. Mean is zero.
}
