% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prior_predictive}
\alias{prior_predictive}
\title{Sample from the prior predictive of a Bayesian Neural Network}
\usage{
prior_predictive(bnn, n = 1)
}
\arguments{
\item{bnn}{BNN obtained using \code{\link{BNN}}}

\item{n}{Number of samples}
}
\value{
matrix of prior predictive samples; Columns are the different samples
}
\description{
Sample from the prior predictive of a Bayesian Neural Network
}
