% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{RNN}
\alias{RNN}
\title{Create a RNN layer with `in_size` input, `out_size` hidden state and `act` activation function}
\usage{
RNN(in_size, out_size, act = c("sigmoid", "tanh", "identity", "relu"))
}
\arguments{
\item{in_size}{Input size}

\item{out_size}{Output size}

\item{act}{Activation function}
}
\description{
Create a RNN layer with `in_size` input, `out_size` hidden state and `act` activation function
}
\seealso{
\code{\link{Dense}}
}
