% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-stepsize.R
\name{sadapter.Const}
\alias{sadapter.Const}
\title{Use a constant stepsize in mcmc}
\usage{
sadapter.Const(l)
}
\arguments{
\item{l}{stepsize}
}
\value{
list with `juliavar`, `juliacode` and the given arguments
}
\description{
Use a constant stepsize in mcmc
}
