% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-stepsize.R
\name{sadapter.DualAverage}
\alias{sadapter.DualAverage}
\title{Use Dual Averaging like in STAN to tune stepsize}
\usage{
sadapter.DualAverage(
  adapt_steps,
  initial_stepsize = 1,
  target_accept = 0.65,
  gamma = 0.05,
  t0 = 10,
  kappa = 0.75
)
}
\arguments{
\item{adapt_steps}{number of adaptation steps}

\item{initial_stepsize}{initial stepsize}

\item{target_accept}{target acceptance ratio}

\item{gamma}{See STAN manual NUTS paper}

\item{t0}{See STAN manual or NUTS paper}

\item{kappa}{See STAN manual or NUTS paper}
}
\value{
list with `juliavar`, `juliacode`, and all given arguments
}
\description{
Use Dual Averaging like in STAN to tune stepsize
}
