% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{sampler.GGMC}
\alias{sampler.GGMC}
\title{Gradient Guided Monte Carlo}
\usage{
sampler.GGMC(
  beta = 0.1,
  l = 1,
  sadapter = sadapter.DualAverage(1000),
  madapter = madapter.FixedMassMatrix(),
  steps = 3
)
}
\arguments{
\item{beta}{See paper}

\item{l}{stepsize}

\item{sadapter}{Stepsize adapter; Not used in original paper}

\item{madapter}{Mass adapter; Not used in ogirinal paper}

\item{steps}{Number of steps before accept/reject}
}
\value{
a list with `juliavar`, `juliacode` and all provided arguments.
}
\description{
Proposed in Garriga-Alonso, A., & Fortuin, V. (2021). Exact langevin dynamics
with stochastic gradients. arXiv preprint arXiv:2102.01691.
}
