% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{sampler.SGLD}
\alias{sampler.SGLD}
\title{Stochastic Gradient Langevin Dynamics as proposed in Welling, M., & Teh, Y. W.
(n.d.). Bayesian Learning via Stochastic Gradient Langevin Dynamics. 8.}
\usage{
sampler.SGLD(
  stepsize_a = 0.1,
  stepsize_b = 0,
  stepsize_gamma = 0.55,
  min_stepsize = -Inf
)
}
\arguments{
\item{stepsize_a}{See eq. above}

\item{stepsize_b}{See eq. above}

\item{stepsize_gamma}{see eq. above}

\item{min_stepsize}{Do not decrease stepsize beyond this}
}
\value{
a list with `juliavar`, `juliacode`, and all given arguments
}
\description{
Stepsizes will be adapted according to
\deqn{a(b+t)^{-\gamma}}
}
