% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{sgld}
\alias{sgld}
\title{Use Stochastic Gradient Longevin Dynamics}
\usage{
sgld(
  bnn,
  batchsize,
  maxiter,
  init = NULL,
  nchains = 1,
  stepsize_a = 0.1,
  stepsize_b = 100,
  stepsize_gamma = 0.8
)
}
\arguments{
\item{bnn}{BNN formed using \code{\link{BNN}}}

\item{batchsize}{Mini-batchsize}

\item{maxiter}{Number of epochs to run SGLD for}

\item{init}{inital values; If `nchains==1` then a numeric vector
of length `BNN.totparams(bnn)`, else a list of
numeric vectors of the size above.}

\item{nchains}{Number of chains to run in parallel.}

\item{stepsize_a}{See the stepsize schedule}

\item{stepsize_b}{See the stepsize schedule}

\item{stepsize_gamma}{See the stepsize schedule}
}
\value{
If `nchains == 1` then return a matrix of dimension
        parameter x draws. Otherwise it returns a tensor of dimension
        parameter x draws x chain
}
\description{
Uses Stochastic Gradient Longevin Dynamics according to
Welling, M., & Teh, Y. W. (2011).
Bayesian learning via stochastic gradient Langevin dynamics.
In Proceedings of the 28th international conference on machine learning (ICML-11) (pp. 681-688).
}
\details{
Stepsizes follow the schedule \eqn{a(b+t)^{-\gamma}}
}
