% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{SIR.laplace}
\alias{SIR.laplace}
\title{Use Sampling-Importance-Resampling to correct laplace approximations}
\usage{
SIR.laplace(bnn, lapprox, n, k)
}
\arguments{
\item{bnn}{A BNN formed using \code{\link{BNN}}}

\item{lapprox}{A laplace approximation formed using \code{\link{laplace}}}

\item{n}{Inital numer of draws}

\item{k}{Final number of re-draws}
}
\value{
A Matrix; Rows are parameters, columns are draws
}
\description{
SIR is implemented using sampling without replacement. Thus, the inital sample
`n` should be much larger than the final sample size `k`
}
