% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BNN}
\alias{summary.BNN}
\title{Print a summary of a BNN}
\usage{
\method{summary}{BNN}(object, ...)
}
\arguments{
\item{object}{A BNN created using \code{\link{BNN}}}

\item{...}{Not used}
}
\description{
Print a summary of a BNN
}
