% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tensor_embed_mat}
\alias{tensor_embed_mat}
\title{Embed a matrix of timeseries into a tensor}
\usage{
tensor_embed_mat(mat, len_seq)
}
\arguments{
\item{mat}{Matrix of time series}

\item{len_seq}{subsequence length}
}
\value{
A tensor of dimension: len_seq x num_vars x num_subsequences
}
\description{
This is used when working with recurrent networks, especially in
the case of seq-to-one modelling. Creates overlapping subsequences
of the data with length `len_seq`. Returned dimensions are seq_len x num_vars x num_subsequences.
}
