% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Truncated}
\alias{Truncated}
\title{Truncates a Distribution}
\usage{
Truncated(dist, lower, upper)
}
\arguments{
\item{dist}{A Julia Distribution created using \code{\link{Gamma}},
\code{\link{InverseGamma}} ...}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
see \code{\link{Gamma}}
}
\description{
Truncates a Julia Distribution between `lower` and `upper`.
}
