% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{vi.get_samples}
\alias{vi.get_samples}
\title{Draw samples form a variational family.}
\usage{
vi.get_samples(vi, n = 1)
}
\arguments{
\item{vi}{obtained using \code{\link{bayes_by_backprop}}}

\item{n}{number of samples}
}
\value{
a matrix whose columns are draws from the variational posterior
}
\description{
Draw samples form a variational family.
}
