% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFluxR-package.R
\docType{package}
\name{BayesFluxR-package}
\alias{BayesFluxR}
\alias{BayesFluxR-package}
\title{BayesFluxR: Implementation of Bayesian Neural Networks}
\description{
Implementation of 'BayesFlux.jl' for R; It extends the famous 'Flux.jl' machine learning library to Bayesian Neural Networks. The goal is not to have the fastest production ready library, but rather to allow more people to be able to use and research on Bayesian Neural Networks.
}
\author{
\strong{Maintainer}: Enrico Wegner \email{e.wegner@student.maastrichtuniversity.nl}

}
\keyword{internal}
