% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFluxR.R
\name{BNN}
\alias{BNN}
\title{Create a Bayesian Neural Network}
\usage{
BNN(x, y, like, prior, init)
}
\arguments{
\item{x}{For a Feedforward structure, this must be a matrix of dimensions
variables x observations; For a recurrent structure, this must be a
tensor of dimensions sequence_length x number_variables x number_sequences;
In general, the last dimension is always the dimension over which will be batched.}

\item{y}{A vector or matrix with observations.}

\item{like}{Likelihood; See for example \code{\link{likelihood.feedforward_normal}}}

\item{prior}{Prior; See for example \code{\link{prior.gaussian}}}

\item{init}{Initialiser; See for example \code{\link{initialise.allsame}}}
}
\value{
List with the following content
\itemize{
    \item `juliavar` - the julia variable containing the BNN
    \item `juliacode` - the string representation of the BNN
}
}
\description{
Create a Bayesian Neural Network
}
