% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-finding.R
\name{find_mode}
\alias{find_mode}
\title{Find the MAP of a BNN using SGD}
\usage{
find_mode(bnn, optimiser, batchsize, epochs)
}
\arguments{
\item{bnn}{a BNN obtained using \code{\link{BNN}}}

\item{optimiser}{an optimiser. These start with `opt.`.
See for example \code{\link{opt.ADAM}}}

\item{batchsize}{batch size}

\item{epochs}{number of epochs to run for}
}
\value{
Returns a vector. Use \code{\link{posterior_predictive}}
to obtain a prediction using this MAP estimate.
}
\description{
Find the MAP of a BNN using SGD
}
