% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{LSTM}
\alias{LSTM}
\title{Create an LSTM layer with `in_size` input size, and `out_size` hidden state size}
\usage{
LSTM(in_size, out_size)
}
\arguments{
\item{in_size}{Input size}

\item{out_size}{Output size}
}
\value{
A list with the following content
\itemize{
     \item in_size - Input Size
     \item out_size - Output Size
     \item julia - Julia code representing the Layer
}
}
\description{
Create an LSTM layer with `in_size` input size, and `out_size` hidden state size
}
\examples{
\dontrun{
  ## Needs previous call to `BayesFluxR_setup` which is time
  ## consuming and requires Julia and BayesFlux.jl
  BayesFluxR_setup(installJulia=TRUE, seed=123)
  net <- Chain(LSTM(5, 5))
}

}
\seealso{
\code{\link{Dense}}
}
